Implementação de controle e gerenciamento de WASocket utilizando a lib baileys

+++ src/app/controllers/TenantController.ts
@@ -9,9 +9,9 @@ class TenantController {
   async index(req: Request, res: Response) {
     try {
       const tenants = await Tenant.find();
-      res.status(200).json(tenants);
+      return res.status(200).json(tenants);
     } catch (error) {
-      res.status(500).json({ error: 'Erro ao listar tenants' });
+      return res.status(500).json({ error: 'Erro ao listar tenants' });
     }
   }
 
@@ -21,14 +21,20 @@ class TenantController {
   async show(req: Request, res: Response) {
     const { id } = req.params;
 
+    if (!id) {
+      return res
+        .status(400)
+        .json({ error: "Id é obrigatória" });
+    }
+
     try {
       const tenant = await Tenant.findById(id);
       if (!tenant) {
         return res.status(404).json({ error: 'Tenant não encontrado' });
       }
-      res.status(200).json(tenant);
+      return res.status(200).json(tenant);
     } catch (error) {
-      res.status(500).json({ error: 'Erro ao obter tenant' });
+      return res.status(500).json({ error: 'Erro ao obter tenant' });
     }
   }
 
@@ -60,9 +66,9 @@ class TenantController {
         usuario,
         senha: hashPwd,
       });
-      res.status(201).json(tenant);
+      return res.status(201).json(tenant);
     } catch (error) {
-      res.status(500).json({ error: 'Erro ao criar tenant' });
+      return res.status(500).json({ error: 'Erro ao criar tenant' });
     }
   }
 
@@ -72,18 +78,22 @@ class TenantController {
   async delete(req: Request, res: Response) {
     const { id } = req.params;
 
+    if (!id) {
+      return res
+        .status(400)
+        .json({ error: "Id é obrigatória" });
+    }
+
     try {
       const deletedTenant = await Tenant.findByIdAndDelete(id);
       if (!deletedTenant) {
         return res.status(404).json({ error: 'Tenant não encontrado' });
       }
-      res.status(200).json({ message: 'Tenant excluído com sucesso' });
+      return res.status(200).json({ message: 'Tenant excluído com sucesso' });
     } catch (error) {
-      res.status(500).json({ error: 'Erro ao excluir tenant' });
+      return res.status(500).json({ error: 'Erro ao excluir tenant' });
     }
   }
-
-
 }
 
 export default new TenantController();
